nilfs2: fix potential out-of-bounds memory access in nilfs_find_entry() (bsc#1235224 CVE-2024-56619).
suse-commit: 67ec84827c565da97e7d5ea8f8c7a49da140f4e7

+++ fs/nilfs2/dir.c
@@ -72,7 +72,7 @@ static inline unsigned long dir_pages(struct inode *inode)
  */
 static unsigned nilfs_last_byte(struct inode *inode, unsigned long page_nr)
 {
-	unsigned last_byte = inode->i_size;
+	u64 last_byte = inode->i_size;
 
 	last_byte -= page_nr << PAGE_CACHE_SHIFT;
 	if (last_byte > PAGE_CACHE_SIZE)
